reorganizing asset loaders, lib/assets contain implementation specific loaders while lib/ contains our internal representation of assets (animations, textures)
