+++ lib/assets/loaders/collada/collada.go
