+++ lib/assets/loaders/collada/collada_test.go
