+++ lib/assets/loaders/collada/newcollada.go
