+++ lib/assets/loaders/collada/rawcollada.go
