+++ lib/assets/loaders/collada/sample/bob.dae
