+++ lib/assets/loaders/collada/sample/cube.dae
