+++ lib/assets/loaders/collada/sample/cube2.dae
