+++ lib/assets/loaders/collada/sample/model.dae
