+++ lib/assets/loaders/collada/sample/model2.dae
