+++ lib/assets/loaders/collada/types.go
