+++ lib/assets/loaders/collada/unfinished.go
