+++ lib/assets/loaders/collada/util.go
