--- lib/types/types.go
@@ -1,5 +1,0 @@
-package types
-
-type Texture struct {
-	ID uint32
-}
